assignment-2-767835/code/partial-pieces/mysimbdp-streamingestmanager/requirements.txt
pika==1.1.0
flask-restful==0.3.5
pymongo
